Added floor texture + player and enemy scripts
Item pick up scripts, changes to FPS scripts, player manager, interactable script, Enemy, Enemy stats
