+++ CPI211/Assets/3rd Party/Hand Painted Stone Texture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2248be380dca134fa884891d890ac41
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
