+++ CPI211/Assets/3rd Party/Hand Painted Stone Texture/Material.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 245a5e4b35b823842bf1dbfeb45fb118
+folderAsset: yes
+timeCreated: 1476631864
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
