+++ CPI211/Assets/3rd Party/Hand Painted Stone Texture/Material/Stone_floor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f73475cf0cad364b9800b85c2a229fc
+timeCreated: 1476631882
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
