+++ CPI211/Assets/3rd Party/Hand Painted Stone Texture/Texture.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 7e8c075e0c1874e4197680e91fbe8e00
+folderAsset: yes
+timeCreated: 1476631859
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
