+++ CPI211/Assets/3rd Party/_TerrainAutoUpgrade.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d09d635de31ef8c4da20d04a42feb84e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
