+++ CPI211/Assets/Materials/glass_mat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26e51be8ad475734baeafa402cd2260b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
