+++ CPI211/Assets/Scripts/Enemy.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Enemy : Interactable
+{
+    public override void Interact()
+    {
+        base.Interact();
+    }
+}
