+++ CPI211/Assets/Scripts/EnemyStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce5e35df52d18744c84543650071a03e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
