+++ CPI211/Assets/Scripts/Interactable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9135ac8bd2afe1b4fa0365d8c0288fbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
