+++ CPI211/Assets/Scripts/ItemPickUP.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ItemPickUP : Interactable
+{
+    public override void Interact()
+    {
+        base.Interact();
+
+        PickUp();
+ 
+    }
+
+    void PickUp()
+    {
+        Debug.Log("Picking Up item");
+    }
+}
