+++ CPI211/Assets/Scripts/ItemPickUP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac179fe4c29a99741aab3234e1bd0d64
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
