--- CPI211/Assets/Scripts/PlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d1fa5d978d156a4cbf7c3e67c06d276
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
