+++ CPI211/Assets/Scripts/PlayerManager.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerManager : MonoBehaviour
+{
+    #region Singleton
+
+    public static PlayerManager instance;
+
+    private void Awake()
+    {
+        instance = this; 
+    }
+
+    #endregion
+
+    public GameObject player;
+
+}
