+++ CPI211/Assets/Scripts/PlayerManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0bdc6471933a9049a37ea3ef13a0222
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
