--- CPI211/Assets/Standard Assets/Characters/FirstPersonCharacter/Scripts/FirstPersonController.cs.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 05ec5cf00ca181d45a42ba1870e148c3
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName:
